move the test to the common file now that android has the same behaviour
