+++ packages/react-native-bpk-component-phone-input/src/BpkDialingCodeList-test.common.js
@@ -21,6 +21,7 @@
 import React from 'react';
 import { Image } from 'react-native';
 import renderer from 'react-test-renderer';
+import BpkSectionList from 'react-native-bpk-component-section-list';
 
 import BpkDialingCodeList from './BpkDialingCodeList';
 
@@ -43,6 +44,11 @@ const CODES = [
     name: 'Two',
   },
 ];
+
+const SUGGESTED = {
+  ids: ['1', '3'],
+  title: 'Suggested for you',
+};
 
 const commonTests = () => {
   jest.mock('Image', () => 'Image');
@@ -73,6 +79,28 @@ const commonTests = () => {
         .toJSON();
       expect(tree).toMatchSnapshot();
     });
+  });
+
+  // This test can be moved to commonTests() once the feature has been implemented on android
+  it('should render correctly with suggested IDs', () => {
+    const tree = renderer.create(
+      <BpkDialingCodeList
+        dialingCodes={[
+          ...CODES,
+          {
+            id: '3',
+            dialingCode: '3',
+            name: 'Three',
+          },
+        ]}
+        onItemPress={jest.fn()}
+        renderFlag={() => null}
+        suggested={SUGGESTED}
+      />,
+    );
+    const { props } = tree.root.findByType(BpkSectionList);
+
+    expect(props.sections).toMatchSnapshot();
   });
 };
